Implement guessing date from filename patterns
If all reliable methods of identifying a timestamp for a photo fail (in
particular the extraction from the metadata json), photos currently
receive the timestamp of the time the run was executed.

This adds a new CLI argument --guess-timestamp-from-filename.
If specified, after trying all other methods of determining the
timestamp, some predefined regex patterns will check the filename for
common date/time patterns, and apply those if found.